pyproject.toml
[project]
name = "thai-receipt-generator"
version = "0.1.0"
description = "Generate varied Thai receipt/invoice images for OCR testing"
requires-python = ">=3.11"
dependencies = [
    "jinja2>=3.1",
    "click>=8.0",
    "playwright>=1.40",
    "python-dotenv>=1.0",
]

[project.scripts]
thai-receipt-gen = "thai_receipt_generator.cli:main"

[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-data]
thai_receipt_generator = ["templates/*.html"]
